# Use optimized DoubleUtil methods