--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/MS/internal/Ink/StrokeNodeData.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/MS/internal/Ink/StrokeNodeData.cs
@@ -58,7 +58,7 @@ namespace MS.Internal.Ink
         { 
             get 
             {
-                Debug.Assert(DoubleUtil.AreClose(0, s_empty._pressure));
+                Debug.Assert(DoubleUtil.IsZero(s_empty._pressure));
                 return DoubleUtil.AreClose(_pressure, s_empty._pressure); 
             } 
         }