--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/MS/internal/Ink/StrokeNodeOperations.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/MS/internal/Ink/StrokeNodeOperations.cs
@@ -80,7 +80,7 @@ namespace MS.Internal.Ink
             System.Diagnostics.Debug.Assert((boundingBox.X <= 0) && (boundingBox.Y <= 0));
 
             double pressureFactor = node.PressureFactor;
-            if (!DoubleUtil.AreClose(pressureFactor,1d))
+            if (!DoubleUtil.IsOne(pressureFactor))
             {
                 boundingBox = new Rect(
                     _shapeBounds.X * pressureFactor,
@@ -97,7 +97,7 @@ namespace MS.Internal.Ink
         internal void GetNodeContourPoints(in StrokeNodeData node, List<Point> pointBuffer)
         {
             double pressureFactor = node.PressureFactor;
-            if (DoubleUtil.AreClose(pressureFactor, 1d))
+            if (DoubleUtil.IsOne(pressureFactor))
             {
                 for (int i = 0; i < _vertices.Length; i++)
                 {
@@ -959,7 +959,7 @@ namespace MS.Internal.Ink
                     // Adjust the arc for the node' pressure factor.
                     Vector hitCenter = hitSegment.Begin + hitSegment.Radius - position;
                     Vector hitRadius = hitSegment.Radius;
-                    if (!DoubleUtil.AreClose(pressureFactor, 1d))
+                    if (!DoubleUtil.IsOne(pressureFactor))
                     {
                         System.Diagnostics.Debug.Assert(DoubleUtil.IsZero(pressureFactor) == false);
                         hitCenter /= pressureFactor;
@@ -984,7 +984,7 @@ namespace MS.Internal.Ink
                     // Adjust the segment for the node's pressure factor
                     Vector hitBegin = hitSegment.Begin - position;
                     Vector hitEnd = hitBegin + hitSegment.Vector;
-                    if (!DoubleUtil.AreClose(pressureFactor, 1d))
+                    if (!DoubleUtil.IsOne(pressureFactor))
                     {
                         System.Diagnostics.Debug.Assert(DoubleUtil.IsZero(pressureFactor) == false);
                         hitBegin /= pressureFactor;