--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/Stylus/Pointer/PointerTouchDevice.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/Stylus/Pointer/PointerTouchDevice.cs
@@ -95,7 +95,7 @@ namespace System.Windows.Input.StylusPointer
 
                 StylusPointPropertyInfo propertyInfo = stylusPoint.Description.GetPropertyInfo(property);
 
-                if (!DoubleUtil.AreClose(propertyInfo.Resolution, 0d))
+                if (!DoubleUtil.IsZero(propertyInfo.Resolution))
                 {
                     value /= propertyInfo.Resolution;
                 }