--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/Stylus/Wisp/WispStylusTouchDevice.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Input/Stylus/Wisp/WispStylusTouchDevice.cs
@@ -56,7 +56,7 @@ namespace System.Windows.Input.StylusWisp
 
                 StylusPointPropertyInfo propertyInfo = stylusPoint.Description.GetPropertyInfo(property);
 
-                if (!DoubleUtil.AreClose(propertyInfo.Resolution, 0d))
+                if (!DoubleUtil.IsZero(propertyInfo.Resolution))
                 {
                     value /= propertyInfo.Resolution;
                 }