--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/UIElement.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/UIElement.cs
@@ -1091,7 +1091,7 @@ namespace System.Windows
             double newValue;
 
             // If DPI == 1, don't use DPI-aware rounding.
-            if (!DoubleUtil.AreClose(dpiScale, 1.0))
+            if (!DoubleUtil.IsOne(dpiScale))
             {
                 newValue = Math.Round(value * dpiScale) / dpiScale;
                 // If rounding produces a value unacceptable to layout (NaN, Infinity or MaxValue), use the original value.