--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Text/Line.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/Text/Line.cs
@@ -172,7 +172,7 @@ namespace MS.Internal.Text
             if (_line.HasOverflowed && _owner.ParagraphProperties.TextTrimming != TextTrimming.None)
             {
                 // We should not shift offset in this case
-                Invariant.Assert(DoubleUtil.AreClose(delta, 0));
+                Invariant.Assert(DoubleUtil.IsZero(delta));
                 System.Windows.Media.TextFormatting.TextLine line = _line.Collapse(GetCollapsingProps(_wrappingWidth, _owner.ParagraphProperties));
                 Invariant.Assert(line.HasCollapsed, "Line has not been collapsed");
                 textBounds = line.GetTextBounds(cp, cch);
@@ -208,7 +208,7 @@ namespace MS.Internal.Text
             if (_line.HasOverflowed && _owner.ParagraphProperties.TextTrimming != TextTrimming.None)
             {
                 System.Windows.Media.TextFormatting.TextLine line = _line.Collapse(GetCollapsingProps(_wrappingWidth, _owner.ParagraphProperties));
-                Invariant.Assert(DoubleUtil.AreClose(delta, 0));
+                Invariant.Assert(DoubleUtil.IsZero(delta));
                 Invariant.Assert(line.HasCollapsed, "Line has not been collapsed");
                 return line.GetCharacterHitFromDistance(distance);
             }
@@ -466,7 +466,7 @@ namespace MS.Internal.Text
             if (_line.HasOverflowed && _owner.ParagraphProperties.TextTrimming != TextTrimming.None)
             {
                 // We should not shift offset in this case
-                Invariant.Assert(DoubleUtil.AreClose(delta, 0));
+                Invariant.Assert(DoubleUtil.IsZero(delta));
                 System.Windows.Media.TextFormatting.TextLine line = _line.Collapse(GetCollapsingProps(_wrappingWidth, _owner.ParagraphProperties));
                 Invariant.Assert(line.HasCollapsed, "Line has not been collapsed");
                 textBounds = line.GetTextBounds(cp, cch);