--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/documents/DocumentGrid.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/MS/Internal/documents/DocumentGrid.cs
@@ -2333,7 +2333,7 @@ namespace MS.Internal.Documents
                 return;
             }
 
-            if (!DoubleUtil.AreClose(1.0, neededScaleFactor))
+            if (!DoubleUtil.IsOne(neededScaleFactor))
             {
                 //Rescale the row.
                 ApplyViewParameters(pivotRow);