--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/DataGrid.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/DataGrid.cs
@@ -1976,7 +1976,7 @@ namespace System.Windows.Controls
             if (!InternalColumns.ColumnWidthsComputationPending)
             {
                 double widthChange = newSize.Width - oldSize.Width;
-                if (!DoubleUtil.AreClose(widthChange, 0.0))
+                if (!DoubleUtil.IsZero(widthChange))
                 {
                     _finalViewportWidth = newSize.Width;
                     if (!_viewportWidthChangeNotificationPending)
@@ -2001,7 +2001,7 @@ namespace System.Windows.Controls
             }
 
             double widthChange = _finalViewportWidth - _originalViewportWidth;
-            if (!DoubleUtil.AreClose(widthChange, 0.0))
+            if (!DoubleUtil.IsZero(widthChange))
             {
                 NotifyPropertyChanged(this,
                     "ViewportWidth",