--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/DataGridCellsPanel.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/DataGridCellsPanel.cs
@@ -126,7 +126,7 @@ namespace System.Windows.Controls
                 childMeasureConstraint.Width = column.GetConstraintWidth(isColumnHeader);
             }
 
-            if (DoubleUtil.AreClose(desiredWidth, 0.0))
+            if (DoubleUtil.IsZero(desiredWidth))
             {
                 child.Measure(childMeasureConstraint);
             }
@@ -140,7 +140,7 @@ namespace System.Windows.Controls
                 // Allow the column to process the desired size
                 column.UpdateDesiredWidthForAutoColumn(
                     isColumnHeader,
-                    DoubleUtil.AreClose(desiredWidth, 0.0) ? childDesiredSize.Width : desiredWidth);
+                    DoubleUtil.IsZero(desiredWidth) ? childDesiredSize.Width : desiredWidth);
 
                 // For auto kind columns measure again with display value if
                 // the desired width is greater than display value.
@@ -250,7 +250,7 @@ namespace System.Windows.Controls
 
             bool hasStarColumns = parentDataGrid.InternalColumns.HasVisibleStarColumns;
             double averageColumnWidth = parentDataGrid.InternalColumns.AverageColumnWidth;
-            bool invalidAverage = DoubleUtil.AreClose(averageColumnWidth, 0.0);
+            bool invalidAverage = DoubleUtil.IsZero(averageColumnWidth);
             bool notVirtualizing = !IsVirtualizing;
             bool generateAll = invalidAverage || hasStarColumns || notVirtualizing;
             int frozenColumnCount = parentDataGrid.FrozenColumnCount;
@@ -357,7 +357,7 @@ namespace System.Windows.Controls
                                     }
 
                                     double cellChoppedWidth = viewportStartX - nextNonFrozenCellStart;
-                                    if (DoubleUtil.AreClose(cellChoppedWidth, 0.0))
+                                    if (DoubleUtil.IsZero(cellChoppedWidth))
                                     {
                                         nextNonFrozenCellStart = nextFrozenCellStart + childSize.Width;
                                         allocatedSpace += childSize.Width;
@@ -1517,7 +1517,7 @@ namespace System.Windows.Controls
                     else
                     {
                         double cellChoppedWidth = arrangeState.ViewportStartX - arrangeState.NextNonFrozenCellStart;
-                        if (DoubleUtil.AreClose(cellChoppedWidth, 0.0))
+                        if (DoubleUtil.IsZero(cellChoppedWidth))
                         {
                             rcChild.X = arrangeState.NextFrozenCellStart;
                             arrangeState.NextNonFrozenCellStart = arrangeState.NextFrozenCellStart + childWidth;
@@ -2021,7 +2021,7 @@ namespace System.Windows.Controls
 
             DataGridRowsPresenter parentRowsPresenter = ParentRowsPresenter;
 
-            if (DoubleUtil.AreClose(availableViewportWidth, 0.0) && parentRowsPresenter != null)
+            if (DoubleUtil.IsZero(availableViewportWidth) && parentRowsPresenter != null)
             {
                 Size rowPresenterAvailableSize = parentRowsPresenter.AvailableSize;
                 if (!double.IsNaN(rowPresenterAvailableSize.Width) && !Double.IsInfinity(rowPresenterAvailableSize.Width))
@@ -2294,7 +2294,7 @@ namespace System.Windows.Controls
                         {
                             columnStart = nextFrozenCellStart;
                             double cellChoppedWidth = viewportStartX - nextNonFrozenCellStart;
-                            if (DoubleUtil.AreClose(cellChoppedWidth, 0.0))
+                            if (DoubleUtil.IsZero(cellChoppedWidth))
                             {
                                 columnEnd = columnStart + columnWidth;
                                 nextNonFrozenCellStart = nextFrozenCellStart + columnWidth;
@@ -2342,7 +2342,7 @@ namespace System.Windows.Controls
                     offsetChange = columnStart - nextFrozenCellStart;
                 }
 
-                if (DoubleUtil.AreClose(offsetChange, 0.0))
+                if (DoubleUtil.IsZero(offsetChange))
                 {
                     return true;
                 }