--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/DataGridColumnCollection.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/DataGridColumnCollection.cs
@@ -811,8 +811,8 @@ namespace System.Windows.Controls
                 if (width.IsStar)
                 {
                     hasStarColumns = true;
-                    if (!DoubleUtil.AreClose(width.Value, 0.0) &&
-                        !DoubleUtil.AreClose(width.DesiredValue, 0.0))
+                    if (!DoubleUtil.IsZero(width.Value) &&
+                        !DoubleUtil.IsZero(width.DesiredValue))
                     {
                         perStarWidth = width.DesiredValue / width.Value;
                         break;
@@ -1615,7 +1615,7 @@ namespace System.Windows.Controls
                 }
 
                 DataGridLength width = column.Width;
-                if (width.IsStar && !DoubleUtil.AreClose(width.Value, 0.0))
+                if (width.IsStar && !DoubleUtil.IsZero(width.Value))
                 {
                     if (DoubleUtil.GreaterThan(width.DisplayValue, column.MinWidth))
                     {
@@ -1835,7 +1835,7 @@ namespace System.Windows.Controls
                         perStarWidth,
                         retainAuto);
 
-                    if (DoubleUtil.AreClose(horizontalChange, 0.0))
+                    if (DoubleUtil.IsZero(horizontalChange))
                     {
                         break;
                     }
@@ -1865,7 +1865,7 @@ namespace System.Windows.Controls
                 }
 
                 DataGridLength width = column.Width;
-                if (width.IsStar && !DoubleUtil.AreClose(width.Value, 0.0))
+                if (width.IsStar && !DoubleUtil.IsZero(width.Value))
                 {
                     if (DoubleUtil.LessThan(width.DisplayValue, column.MaxWidth))
                     {