--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/ScrollViewer.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/ScrollViewer.cs
@@ -1624,13 +1624,13 @@ namespace System.Windows.Controls
                     double viewportHeight = ViewportHeight + 1d; // Using +1 to account for last partially visible item in viewport
                     if (viewport != null)
                     {
-                        _panningInfo.DeltaPerHorizontalOffet = (DoubleUtil.AreClose(viewportWidth, 0) ? 0 : viewport.ActualWidth / viewportWidth);
-                        _panningInfo.DeltaPerVerticalOffset = (DoubleUtil.AreClose(viewportHeight, 0) ? 0 : viewport.ActualHeight / viewportHeight);
+                        _panningInfo.DeltaPerHorizontalOffet = (DoubleUtil.IsZero(viewportWidth) ? 0 : viewport.ActualWidth / viewportWidth);
+                        _panningInfo.DeltaPerVerticalOffset = (DoubleUtil.IsZero(viewportHeight) ? 0 : viewport.ActualHeight / viewportHeight);
                     }
                     else
                     {
-                        _panningInfo.DeltaPerHorizontalOffet = (DoubleUtil.AreClose(viewportWidth, 0) ? 0 : ActualWidth / viewportWidth);
-                        _panningInfo.DeltaPerVerticalOffset = (DoubleUtil.AreClose(viewportHeight, 0) ? 0 : ActualHeight / viewportHeight);
+                        _panningInfo.DeltaPerHorizontalOffet = (DoubleUtil.IsZero(viewportWidth) ? 0 : ActualWidth / viewportWidth);
+                        _panningInfo.DeltaPerVerticalOffset = (DoubleUtil.IsZero(viewportHeight) ? 0 : ActualHeight / viewportHeight);
                     }
 
                     // Template bind other Scroll Manipulation properties if needed.
@@ -1758,7 +1758,7 @@ namespace System.Windows.Controls
                 {
                     unusedX = 0;
                 }
-                _panningInfo.InHorizontalFeedback = (!DoubleUtil.AreClose(unusedX, 0));
+                _panningInfo.InHorizontalFeedback = (!DoubleUtil.IsZero(unusedX));
 
                 double unusedY = _panningInfo.UnusedTranslation.Y;
                 if (!_panningInfo.InVerticalFeedback &&
@@ -1766,7 +1766,7 @@ namespace System.Windows.Controls
                 {
                     unusedY = 0;
                 }
-                _panningInfo.InVerticalFeedback = (!DoubleUtil.AreClose(unusedY, 0));
+                _panningInfo.InVerticalFeedback = (!DoubleUtil.IsZero(unusedY));
 
                 if (_panningInfo.InHorizontalFeedback || _panningInfo.InVerticalFeedback)
                 {
@@ -1787,7 +1787,7 @@ namespace System.Windows.Controls
             double offset = (isHorizontal ? HorizontalOffset : VerticalOffset);
             double scrollableLength = (isHorizontal ? ScrollableWidth : ScrollableHeight);
 
-            if (DoubleUtil.AreClose(scrollableLength, 0))
+            if (DoubleUtil.IsZero(scrollableLength))
             {
                 // If the Scrollable length in this direction is 0,
                 // then we should neither scroll nor report the boundary feedback
@@ -1823,7 +1823,7 @@ namespace System.Windows.Controls
 
             if (isHorizontal)
             {
-                if (!DoubleUtil.AreClose(delta, 0))
+                if (!DoubleUtil.IsZero(delta))
                 {
                     // if there is any delta left, then re-evalute the horizontal offset
                     ScrollToHorizontalOffset(_panningInfo.OriginalHorizontalOffset -
@@ -1833,7 +1833,7 @@ namespace System.Windows.Controls
             }
             else
             {
-                if (!DoubleUtil.AreClose(delta, 0))
+                if (!DoubleUtil.IsZero(delta))
                 {
                     // if there is any delta left, then re-evalute the vertical offset
                     ScrollToVerticalOffset(_panningInfo.OriginalVerticalOffset -