--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/Slider.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/Slider.cs
@@ -1009,7 +1009,7 @@ namespace System.Windows.Controls
             if (Orientation == Orientation.Horizontal)
             {
                 // Calculate part size for HorizontalSlider
-                if (DoubleUtil.AreClose(range, 0d) || (DoubleUtil.AreClose(trackSize.Width, thumbSize.Width)))
+                if (DoubleUtil.IsZero(range) || (DoubleUtil.AreClose(trackSize.Width, thumbSize.Width)))
                 {
                     valueToSize = 0d;
                 }
@@ -1031,7 +1031,7 @@ namespace System.Windows.Controls
             else
             {
                 // Calculate part size for VerticalSlider
-                if (DoubleUtil.AreClose(range, 0d) || (DoubleUtil.AreClose(trackSize.Height, thumbSize.Height)))
+                if (DoubleUtil.IsZero(range) || (DoubleUtil.AreClose(trackSize.Height, thumbSize.Height)))
                 {
                     valueToSize = 0d;
                 }
@@ -1074,7 +1074,7 @@ namespace System.Windows.Controls
             if (Orientation == Orientation.Horizontal)
             {
                 // Calculate part size for HorizontalSlider
-                if (DoubleUtil.AreClose(range, 0d) || (DoubleUtil.AreClose(trackSize.Width, thumbSize.Width)))
+                if (DoubleUtil.IsZero(range) || (DoubleUtil.AreClose(trackSize.Width, thumbSize.Width)))
                 {
                     valueToSize = 0d;
                 }
@@ -1098,7 +1098,7 @@ namespace System.Windows.Controls
             else
             {
                 // Calculate part size for VerticalSlider
-                if (DoubleUtil.AreClose(range, 0d) || (DoubleUtil.AreClose(trackSize.Height, thumbSize.Height)))
+                if (DoubleUtil.IsZero(range) || (DoubleUtil.AreClose(trackSize.Height, thumbSize.Height)))
                 {
                     valueToSize = 0d;
                 }