--- a/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Markup/XamlFigureLengthSerializer.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Markup/XamlFigureLengthSerializer.cs
@@ -245,7 +245,7 @@ namespace System.Windows.Markup
             else
             {
                 Debug.Assert(   unit == FigureUnitType.Pixel 
-                            ||  DoubleUtil.AreClose(unitFactor, 1.0)    );
+                            ||  DoubleUtil.IsOne(unitFactor)    );
 
                 ReadOnlySpan<char> valueString = valueSpan.Slice(0, valueSpan.Length - strLenUnit);
                 value = double.Parse(valueString, provider: cultureInfo) * unitFactor;