## Description
Uses existing optimized DoubleUtil methods instead of DoubleUtil.AreClose. It's a bit cleaner and provides tiny improvements in speed, IL size and JIT code size. The improvements are negligible but it's overall better so I believe the changes are worth it.

<details>
  <summary>Benchmark results for DoubleUtil.AreClose vs DoubleUtil.IsZero</summary>
  
| Method   | value                | Mean      | Error     | StdDev    | Ratio | RatioSD | Code Size |
|--------- |--------------------- |----------:|----------:|----------:|------:|--------:|----------:|
| AreClose | 1.110(...)5E-16 [22] | 0.2316 ns | 0.0034 ns | 0.0026 ns |  1.00 |    0.02 |      80 B |
| IsZero   | 1.110(...)5E-16 [22] | 0.0235 ns | 0.0099 ns | 0.0093 ns |  0.10 |    0.04 |      27 B |
|          |                      |           |           |           |       |         |           |
| AreClose | 1                    | 0.2294 ns | 0.0065 ns | 0.0061 ns |  1.00 |    0.04 |      80 B |
| IsZero   | 1                    | 0.0133 ns | 0.0044 ns | 0.0037 ns |  0.06 |    0.02 |      27 B |
  
</details>

<details>
  <summary>Benchmark results for DoubleUtil.AreClose vs DoubleUtil.IsOne</summary>
  
| Method   | value | Mean      | Error     | StdDev    | Ratio | RatioSD | Code Size |
|--------- |------ |----------:|----------:|----------:|------:|--------:|----------:|
| AreClose | 1     | 0.2084 ns | 0.0116 ns | 0.0103 ns |  1.00 |    0.07 |      88 B |
| IsOne    | 1     | 0.0142 ns | 0.0071 ns | 0.0059 ns |  0.07 |    0.03 |      35 B |
|          |       |           |           |           |       |         |           |
| AreClose | 2     | 0.2385 ns | 0.0058 ns | 0.0049 ns |  1.00 |    0.03 |      88 B |
| IsOne    | 2     | 0.0217 ns | 0.0069 ns | 0.0064 ns |  0.09 |    0.03 |      35 B |
  
</details>

## Customer Impact
Tiny improvements in speed, IL size and JIT code size.

## Regression
No.

## Testing
Local testing.

## Risk
Low.